return type y

+++ main_spark.py
@@ -51,7 +51,7 @@ def main(sc):
     pos_neg_score = udf(lambda b_id: MLVectors.dense(get_PosNeg_score(b_id)), VectorUDT())
     # clarity = udf(lambda b_id: get_clarity(b_id), ArrayType(FloatType()))
     elite_cnt = udf(lambda b_id: get_elite_cnt(b_id), IntegerType())
-    label = udf(lambda b_id: get_y(b_id), FloatType())
+    label = udf(lambda b_id: get_y(b_id), IntegerType())
 
     # Generate feature columns
     # data_f = data_f.withColumn("city", city(data_f['biz_id']))
@@ -86,9 +86,9 @@ def main(sc):
 
     train_dd = (train_d.select(col("y"), col("features")) \
                 .rdd \
-                .map(lambda row: LabeledPoint(row.y, MLLibVectors.fromML(row.features))))
+                .map(lambda row: LabeledPoint(float(row.y), MLLibVectors.fromML(row.features))))
     m = SVMWithSGD.train(train_dd)
-    predictionAndLabels = train_d.rdd.map(lambda lp: (float(m.predict(MLLibVectors.fromML(lp.features))), lp.y))
+    predictionAndLabels = train_d.rdd.map(lambda lp: (float(m.predict(MLLibVectors.fromML(lp.features))), float(lp.y)))
     # Grid search for best params and model
     # scores = {}
     # max_score = 0
